engine/sdk/py-kit-engine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hop-top-kit-engine"
version = "0.1.0"
description = "Python SDK for kit serve — identity, storage, sync, peers via sidecar"
requires-python = ">=3.10"
dependencies = ["requests>=2.28"]

[project.optional-dependencies]
ws = ["websocket-client>=1.6"]
dev = ["pytest>=8", "responses>=0.25", "ruff>=0.8"]

[tool.hatch.build.targets.wheel]
packages = ["kit_engine"]
